+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Improve the UI design, changes names and delete prints for debugging
 - Added test structure to README file
+- Improved menu design to be more responsive
 
 ### Fixed
 - N/A